requirements.txt
streamlit>=1.35.0
slack-sdk>=3.27.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
schedule>=1.2.0
